requirements.txt
argparse==1.2.1
boto==2.27.0
gnupg==1.4.0
nose==1.2.1
path.py==2.4.1
pep8==1.5.7
pyPdf2==1.23
python-bidi==0.3.4
PyYAML==3.11
reportlab==3.1.44
requests==2.3.0
git+https://github.com/edx/opaque-keys.git@1254ed4d615a428591850656f39f26509b86d30a#egg=opaque-keys

# i18n packages
Babel==1.3
transifex-client==0.12.1
SQLAlchemy==1.1.12
